+++ isitfit/cost/mainManager.py
@@ -32,9 +32,10 @@ class MainManager:
         # click context for errors
         self.ctx = ctx
 
+
+    def set_iterator(self, ec2_it):
         # generic iterator (iterates over regions and items)
-        from isitfit.cost.redshift.iterator import Ec2Iterator
-        self.ec2_it = Ec2Iterator()
+        self.ec2_it = ec2_it
 
 
     def add_listener(self, event, listener):
@@ -46,59 +47,16 @@ class MainManager:
       self.listeners[event].append(listener)
 
 
-    def ec2_count(self):
-      # method 1
-      # ec2_it = self.ec2_resource.instances.all()
-      # return len(list(ec2_it))
-
-      # method 2, using the generic iterator
-      nc = len(list(self.ec2_it.iterate_core(True, True)))
-      msg_count = "Found a total of %i EC2 instance(s) in %i region(s) (other regions do not hold any EC2)"
-      logger.warning(msg_count%(nc, len(self.ec2_it.region_include)))
-      return nc
-
-
-    def ec2_iterator(self):
-      # method 1
-      # ec2_it = self.ec2_resource.instances.all()
-      # return ec2_it
-
-      # boto3 ec2 and cloudwatch data
-      ec2_resource_all = {}
-
-      # TODO cannot use directly use the iterator exposed in "ec2_it"
-      # because it would return the dataframes from Cloudwatch,
-      # whereas in the cloudwatch data fetch here, the data gets cached to redis.
-      # Once the redshift.iterator can cache to redis, then the cloudwatch part here
-      # can also be dropped, as well as using the "ec2_it" iterator directly
-      # for ec2_dict in self.ec2_it:
-      for ec2_dict in self.ec2_it.iterate_core(True, False):
-        if ec2_dict['Region'] not in ec2_resource_all.keys():
-          boto3.setup_default_session(region_name = ec2_dict['Region'])
-          ec2_resource_all[ec2_dict['Region']] = boto3.resource('ec2')
-
-        ec2_resource_single = ec2_resource_all[ec2_dict['Region']]
-        ec2_l = ec2_resource_single.instances.filter(InstanceIds=[ec2_dict['InstanceId']])
-        ec2_l = list(ec2_l)
-        if len(ec2_l)==0:
-          continue # not found
-
-        # yield first entry
-        ec2_obj = ec2_l[0]
-        ec2_obj.region_name = ec2_dict['Region']
-        yield ec2_obj
-
-
     def get_ifi(self):
         # 0th pass to count
-        n_ec2_total = self.ec2_count()
+        n_ec2_total = self.ec2_it.count()
 
         if n_ec2_total==0:
           return
 
         # context for pre listeners
         context_pre = {}
-        context_pre['ec2_instances'] = self.ec2_iterator()
+        context_pre['ec2_instances'] = self.ec2_it
         context_pre['region_include'] = self.ec2_it.region_include
         context_pre['n_ec2_total'] = n_ec2_total
         context_pre['click_ctx'] = self.ctx
@@ -117,7 +75,8 @@ class MainManager:
         ec2_noCloudwatch = []
         ec2_noCloudtrail = []
         # Edit 2019-11-12 use "initial=0" instead of "=1". Check more details in a similar note in "cloudtrail_ec2type.py"
-        for ec2_obj in tqdm(self.ec2_iterator(), total=n_ec2_total, desc="Pass 2/2 through EC2 instances", initial=0):
+        iter_wrap = tqdm(self.ec2_it, total=n_ec2_total, desc="Pass 2/2 through %s"%self.ec2_it.service_description, initial=0)
+        for ec2_obj in iter_wrap:
 
           try:
             # context dict to be passed between listeners