+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Version latest (0.15.0rc?, 2019-09-24?)
 - enh: move `cloudtrail` manager out of `mainManger` to listeners
 - enh: move `cache_man` to listener
 - enh: move `ec2_catalog` and create `ec2_common` out of the code in `mainManager` which isn't really part of the workflow
+- enh: move iterator out of mainManager to cli.cost
 
 
 Version 0.14.0 (2019-11-15)